Step 4: TBSS Part 2
The fourth script to run is Step4_RunTBSS_Pt2.sh. This performs the second step of the TBSS processing pipeline on the FA data, tbss_2_reg -T, which applies nonlinear registration to all of the FA images. You only need to change the file directory to your TBSS output location.