Step 6: Run TBSS Part 4
The sixth script to run is Step6_RunTBSS_Pt4.sh. This performs the fourth and final step of the TBSS processing pipeline on the FA data, tbss_4_prestats 0.2, which simply thresholds the skeletonized images as is standard in the field. You only need to change the file directory to your TBSS output location. Should run within a few seconds.