Step 12: Statistical Analyses in R
The twelfth and final script to run is Step12_StatisticalAnalysis.Rmd. The first thing to do is save the output from Step XI at the same location as specified on line 24 of this script. Thereafter the first chunk simply loads in all necessary data and libraries similar to the Step 8 script. For the Addiction Biology manuscript as of July 2024 (before revisions made), Chunk II provides the descriptive and inferential statistics necessary to fill tables 1 and 2; Chunk III provides the descriptive and inferential statistic necessary to fill table 3, as well as the charts shown in Figure 1 panel, D-F.